python/requirements.txt
# Python dependencies for Weather Aggregator
aiohttp>=3.9.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
